training/requirements.txt
torch>=2.3.0
transformers>=4.46.0
datasets>=3.0.0
peft>=0.13.0
trl>=0.11.0
bitsandbytes>=0.44.0
accelerate>=1.0.0
sentencepiece>=0.2.0
tokenizers>=0.20.0
wandb>=0.18.0
scipy>=1.13.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
pyyaml>=6.0.0
